% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_layers.R
\name{keras_custom_model_template}
\alias{keras_custom_model_template}
\title{Generate Keras Custo Model Template}
\usage{
keras_custom_model_template()
}
\description{
Generate Keras Custo Model Template
}
