% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_layer_to_model.R
\name{keras_layer_to_model}
\alias{keras_layer_to_model}
\title{Convert a Keras layer quickly to a custom Keras model so it can be used in eager mode.}
\usage{
keras_layer_to_model(layer, build_model = FALSE)
}
\arguments{
\item{layer}{Layer to convert to a custom model. Should be an expression, where the layer is
invoked with desired parameter values.}

\item{build_model}{Should the built custom model be returned? Or a function that can be used to build
the model? Default FALSE.}
}
\value{
Either a layer or a function to build the layer.
}
\description{
Mainly used for testing new custom Keras layers quickly and easily by converting them to
custom Keras models, which can be used in tensorflow eager mode
}
\details{
Don't forget to set \code{use_implementation("tensorflow")}!
}
