% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_layers.R
\name{layer_instance_norm_params_as_input}
\alias{layer_instance_norm_params_as_input}
\title{Instance Normalization Layer where the Parameters are passed in as inputs}
\usage{
layer_instance_norm_params_as_input(object, output_params = TRUE,
  epsilon = 0.001, name = NULL, trainable = TRUE)
}
\arguments{
\item{object}{Named list of input tensors. Should be named "images", "beta", and "gamma", for the image
tensor, and the beta and gamma affine parameters respectively.}

\item{output_params}{Should the affine parameters be outputted as well? Or just the normalized image? Default: TRUE.}

\item{epsilon}{epsilon value. Small value to add for numerical stability.}

\item{name}{Optional layer name}

\item{trainable}{Is the layer trainable?}
}
\value{
A list of output tensors, or a single tensor.
}
\description{
This layer performs an instance normalization using parameters passed in the inputs. This is
useful if you want to calculate the affine transformation parameters somewhere else in the network,
so this should keep the backpropagation working fin.
}
