% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_layers.R
\name{layer_subpixel_conv2d}
\alias{layer_subpixel_conv2d}
\title{Keras layer to do a subpixel CNN, upsampling layer}
\usage{
layer_subpixel_conv2d(object, scale = 2, name = NULL,
  trainable = TRUE)
}
\arguments{
\item{object}{A tensor input to the layer}

\item{scale}{Upsample by this much (2 = 2x)}

\item{name}{Optional name for the layer}

\item{trainable}{Is the layer trainable?}
}
\value{
A tensor output
}
\description{
This layer does upsampling using a subpixel CNN, sometimes called a pixel shuffle. Essentially it shuffles
together channels such that each new pixel created uses one of the channels in for that pixel in the previous
layer.
}
